requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
mongoengine==0.23
crispy-bootstrap5
django-allauth
djangorestframework>=3.1,<4.0
django-rest-framework-mongoengine>=3.4,<5
blinker==1.4
djangorestframework-simplejwt>=4.6,<5.0
